+++ packages/dara-core/tests/python/test_data_variable.py
@@ -115,15 +115,14 @@ async def test_fetching_global_data_variable():
         assert response.json() == 5
 
         # Check that schema can be fetched
-        response = await client.get('/api/core/data-variable/uid/schema', json={}, headers=AUTH_HEADERS)
-        assert response.status_code == 200
+        response = await client.get('/api/core/data-variable/uid/schema', headers=AUTH_HEADERS)
+        assert response.status_code == 200, response.text
         data = response.json()
         assert {'name': '__col__1__col1', 'type': 'integer'} in data['fields']
         assert {'name': '__col__2__col2', 'type': 'integer'} in data['fields']
         assert {'name': '__col__3__col3', 'type': 'string'} in data['fields']
         assert {'name': '__col__4__col4', 'type': 'string'} in data['fields']
         assert {'name': '__index__0__index', 'type': 'integer'} in data['fields']
-            
 
 
 async def test_fetching_global_data_variable_filters():