+++ packages/dara-core/tests/python/test_derived_data_variable.py
@@ -297,6 +297,19 @@ async def test_derived_data_variable_no_filters():
         assert count_response.json() == 5
         assert mock_func.call_count == 2   # not called again
 
+        # Check schema can be returned correctly
+        scshema_response = await client.get(
+            '/api/core/data-variable/uid/schema?cache_key=' + third_response.json()['cache_key'],
+            headers=AUTH_HEADERS,
+        )
+        assert scshema_response.status_code == 200, scshema_response.text
+        schema = scshema_response.json()
+        assert {'name': '__col__1__col1', 'type': 'integer'} in schema['fields']
+        assert {'name': '__col__2__col2', 'type': 'integer'} in schema['fields']
+        assert {'name': '__col__3__col3', 'type': 'string'} in schema['fields']
+        assert {'name': '__col__4__col4', 'type': 'string'} in schema['fields']
+        assert {'name': '__index__0__index', 'type': 'integer'} in schema['fields']
+
 
 async def test_derived_data_variable_with_filters():
     builder = ConfigurationBuilder()
@@ -343,7 +356,9 @@ async def test_derived_data_variable_with_filters():
             json={'filters': {'column': 'col1', 'value': 2}, 'cache_key': cache_key, 'ws_channel': 'test_channel'},
             headers=AUTH_HEADERS,
         )
-        assert data_second_response.json() == df_convert_to_internal(unfiltered[unfiltered['col1'] == 2]).to_dict(orient='records')
+        assert data_second_response.json() == df_convert_to_internal(unfiltered[unfiltered['col1'] == 2]).to_dict(
+            orient='records'
+        )
         assert mock_func.call_count == 1
 
         # Check count cannot be returned for different filters